chore: rename docs folder and update imports
