+++ backend/src/models/curriculum/curriculums.model.js
@@ -1,5 +1,5 @@
 const db = require('@database/database');
-const flattenSections = require('@utils/flatten.js');
+const { flattenSections } = require('../../../../docs/shared/utils/flatten.js');
 const Curriculum = require('./curriculum.model');
