+++ docs/types/objects/section.js
