+++ docs/types/objects/suggestion.js
