+++ frontend/jsconfig.json
@@ -17,7 +17,7 @@
             "@hooks/*": ["src/hooks/*"],
             "@sections/*": ["src/pages/Home/sections/*"],
             "@lib/*": ["src/lib/*"],
-            "@documentation/*": ["../documentation/*"]
+            "@docs/*": ["../docs/*"]
         },
         "jsx": "react-jsx",
         "checkJs": true,
@@ -27,5 +27,5 @@
         "forceConsistentCasingInFileNames": true
     },
     "exclude": ["node_modules", "dist"],
-    "include": ["src", "../documentation"]
+    "include": ["src", "../docs"]
 }
