+++ frontend/src/features/editor-in-chief/useEditorInChief.jsx
@@ -8,7 +8,7 @@ import {
 import { UserContext } from '@context/UserProvider';
 import { logger } from '@utils/logger';
 const log = logger.create('useEditorInChief.js');
-import { EditorInChiefActions } from '@documentation/constants/actions';
+import { EditorInChiefActions } from '@docs/constants/actions';
 import { postVersion } from '@utils/api-handlers/curriculums/post-version';
 
 export const EditorInChief = EditorInChiefActions;
